# Anonymous-Proxy-Configurator/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MultiProxy Anonymous Router API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface SystemStatus {
  torInstalled: boolean;
  torRunning: boolean;
  proxychainsInstalled: boolean;
  proxychainsConfigured: boolean;
  proxyCount: number;
  activeProxyCount: number;
  torVersion?: string;
  proxychainsVersion?: string;
  uptime?: string;
  lastChecked: string;
}

export interface SystemInfo {
  distro: string;
  distroVersion?: string;
  kernel?: string;
  cpuUsagePercent: number;
  ramUsedMb: number;
  ramTotalMb: number;
  ramUsagePercent: number;
  uptime?: string;
  hostname?: string;
}

export interface AppConfig {
  torPort: number;
  torControlPort: number;
  rotationIntervalSeconds: number;
  autoRemoveDeadProxies: boolean;
  logFilePath?: string;
  proxychainsConfigPath?: string;
  maxRetries?: number;
}

export interface RotationStatus {
  active: boolean;
  intervalSeconds?: number;
  /** Seconds until next rotation */
  nextRotationIn?: number;
  rotationCount: number;
  currentProxyId?: number;
}

export interface InstallResult {
  success: boolean;
  message: string;
  alreadyInstalled: boolean;
  version?: string;
}

export interface ServiceResult {
  success: boolean;
  message: string;
  status: string;
}

export interface ConfigResult {
  success: boolean;
  message: string;
  configPath?: string;
  proxiesConfigured?: number;
}

export interface BulkUploadResult {
  added: number;
  skipped: number;
  errors: number;
  total: number;
  message: string;
}

export type ProxyType = (typeof ProxyType)[keyof typeof ProxyType];

export const ProxyType = {
  socks5: "socks5",
  socks4: "socks4",
  http: "http",
  https: "https",
} as const;

export type ProxyStatus = (typeof ProxyStatus)[keyof typeof ProxyStatus];

export const ProxyStatus = {
  unknown: "unknown",
  alive: "alive",
  dead: "dead",
  checking: "checking",
} as const;

export type ProxyRating = (typeof ProxyRating)[keyof typeof ProxyRating];

export const ProxyRating = {
  fast: "fast",
  medium: "medium",
  slow: "slow",
  unknown: "unknown",
} as const;

export interface Proxy {
  id: number;
  type: ProxyType;
  host: string;
  port: number;
  username?: string;
  password?: string;
  enabled: boolean;
  status: ProxyStatus;
  latencyMs?: number;
  rating: ProxyRating;
  lastChecked?: string;
  createdAt: string;
}

export type CreateProxyType =
  (typeof CreateProxyType)[keyof typeof CreateProxyType];

export const CreateProxyType = {
  socks5: "socks5",
  socks4: "socks4",
  http: "http",
  https: "https",
} as const;

export interface CreateProxy {
  type: CreateProxyType;
  host: string;
  port: number;
  username?: string;
  password?: string;
  enabled?: boolean;
}

export type ProxyValidationResultRating =
  (typeof ProxyValidationResultRating)[keyof typeof ProxyValidationResultRating];

export const ProxyValidationResultRating = {
  fast: "fast",
  medium: "medium",
  slow: "slow",
  unknown: "unknown",
} as const;

export interface ProxyValidationResult {
  proxyId: number;
  host: string;
  port: number;
  alive: boolean;
  latencyMs?: number;
  rating: ProxyValidationResultRating;
  message?: string;
}

export interface IpInfo {
  ip?: string;
  country?: string;
  countryCode?: string;
  city?: string;
  region?: string;
  isp?: string;
  org?: string;
  timezone?: string;
  isTor?: boolean;
  success: boolean;
  error?: string;
}

export interface LeakTestResult {
  leaked: boolean;
  dnsServers: string[];
  message?: string;
  details?: string;
  testedAt: string;
}

export type LogEntryLevel = (typeof LogEntryLevel)[keyof typeof LogEntryLevel];

export const LogEntryLevel = {
  info: "info",
  warn: "warn",
  error: "error",
} as const;

export interface LogEntry {
  id: number;
  level: LogEntryLevel;
  message: string;
  source: string;
  timestamp: string;
  metadata?: string;
}

export interface DeleteResult {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiError {
  error: string;
}

export type StartProxyRotationBody = {
  /** Rotation interval in seconds */
  intervalSeconds?: number;
};

export type BulkUploadProxiesBodyDefaultType =
  (typeof BulkUploadProxiesBodyDefaultType)[keyof typeof BulkUploadProxiesBodyDefaultType];

export const BulkUploadProxiesBodyDefaultType = {
  socks5: "socks5",
  socks4: "socks4",
  http: "http",
  https: "https",
} as const;

export type BulkUploadProxiesBody = {
  /** Newline-separated list of proxies (format: type:host:port or host:port) */
  text: string;
  defaultType?: BulkUploadProxiesBodyDefaultType;
};

export type GetLogsParams = {
  limit?: number;
  level?: GetLogsLevel;
};

export type GetLogsLevel = (typeof GetLogsLevel)[keyof typeof GetLogsLevel];

export const GetLogsLevel = {
  info: "info",
  warn: "warn",
  error: "error",
  all: "all",
} as const;